fix(components): standardize event handling for updateFrontmatter
Unify the event handler pattern across all item components by:
1. Adding loading prop to ItemDebt component
2. Moving form schema from ItemDebt to Detailed component
3. Standardizing event emission with payload parameter
